\name{persDx-package}
\alias{persDx-package}
\docType{package}
\title{Estimating Personalized Diagnostics Rules}
\description{Recommend the optimal biomarker (or test) for disease screening or diagnosis based on patients' individual characteristics.}
\details{
\tabular{ll}{
Package: \tab persDx\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2022-12-13\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung [aut,cre]
Maintainer: Yunro Chung <yunro.chung@asu.edu>
}
\references{Yaliang Zhang and Yunro Chung, Estimating individualized diagnostics rules via linear grid search with applications to subgroup-specific biomarker discovery (in progress)}
\concept{Subgroup identification}
\concept{Receiver operating characteristics curve}
\concept{Sensitivity}
\concept{Specificity}
\concept{Personalized medicine}
