\name{persDx-package}
\alias{persDx-package}
\docType{package}
\title{Estimating Personalized Diagnostics Rules}
\description{Personalized Recommendation of biomarkers or screening/diagnostic tests based on patients' individual profile.}
\details{
\tabular{ll}{
Package: \tab persDx\cr
Type: \tab Package\cr
Version: \tab 0.5.0\cr
Date: \tab 2023-08-15\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung [aut,cre]
Maintainer: Yunro Chung <yunro.chung@asu.edu>
}
\references{Yaliang Zhang and Yunro Chung, Nonparametric estimation of linear personalized diagnostics rules via efficient grid algorithm (under revision)}
\concept{Subgroup identification}
\concept{Receiver operating characteristics curve}
\concept{Area under the ROC curve}
\concept{Survival analysis}
\concept{Personalized medicine}
\concept{Precision medicine}

